Using RunningStat and removed unused code to keep things simpler.
By using RunningStat.min to report the performance as it
give more consistent results because of the high standard
deviation that is generally seen.

Also increased the number of loops for determing the overhead
values (os) so the cpu is "warmed up", which again gives more
consistent results.

Finally the osLoops and msLoops should probably be determined
dynamically.